Update Implement Service

+++ ecommerce/src/main/java/nashtech/phucldh/ecommerce/sercurity/service/impl/UserDetailsImpl.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import nashtech.phucldh.ecommerce.entity.Account;
@@ -14,140 +15,141 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 public class UserDetailsImpl implements UserDetails {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	private Long id;
+    private Long id;
 
-	private String username;
+    private String username;
 
-	@JsonIgnore
-	private String password;
+    @JsonIgnore
+    private String password;
 
-	private String fullname;
+    private String fullname;
 
-	private String email;
+    private String email;
 
-	private int status;
+    private Long status;
 
-	private Collection<? extends GrantedAuthority> authorities;
+    private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl() {
-	}
+    public UserDetailsImpl() {
+    }
 
-	public UserDetailsImpl(Long id, String username, String password, String fullname, String email,
-			int status, Collection<? extends GrantedAuthority> authorities) {
-		this.id = id;
-		this.username = username;
-		this.password = password;
-		this.fullname = fullname;
-		this.email = email;
-		this.status = status;
-		this.authorities = authorities;
-	}
+    public UserDetailsImpl(Long id, String username, String password, String fullname, String email,
+                           Long status, Collection<? extends GrantedAuthority> authorities) {
+        this.id = id;
+        this.username = username;
+        this.password = password;
+        this.fullname = fullname;
+        this.email = email;
+        this.status = status;
+        this.authorities = authorities;
+    }
 
-	public static UserDetailsImpl build(Account theAccount) {
-		List<GrantedAuthority> authorities = theAccount.getRoles().stream()
-				.map(role -> new SimpleGrantedAuthority(role.getName().name())).collect(Collectors.toList());
-		return new UserDetailsImpl(theAccount.getId(), theAccount.getUsername(), theAccount.getPassword(),
-				theAccount.getFullname(), theAccount.getEmail(), theAccount.getStatus(), authorities);
-	}
+    public static UserDetailsImpl build(Account theAccount) {
+        List<GrantedAuthority> authorities = theAccount.getRoles().stream()
+                .map(role -> new SimpleGrantedAuthority(role.getName().name())).collect(Collectors.toList());
+        return new UserDetailsImpl(theAccount.getId(), theAccount.getUsername(), theAccount.getPassword(),
+                theAccount.getFullname(), theAccount.getEmail(), theAccount.getStatus(), authorities);
+    }
 
-	public Long getId() {
-		return id;
-	}
+    public Long getId() {
+        return id;
+    }
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public String getUsername() {
-		return username;
-	}
+    public String getUsername() {
+        return username;
+    }
 
-	public void setUsername(String username) {
-		this.username = username;
-	}
+    public void setUsername(String username) {
+        this.username = username;
+    }
 
-	public String getPassword() {
-		return password;
-	}
+    public String getPassword() {
+        return password;
+    }
 
-	public void setPassword(String password) {
-		this.password = password;
-	}
+    public void setPassword(String password) {
+        this.password = password;
+    }
 
-	public String getFullname() {
-		return fullname;
-	}
+    public String getFullname() {
+        return fullname;
+    }
 
-	public void setFullname(String fullname) {
-		this.fullname = fullname;
-	}
+    public void setFullname(String fullname) {
+        this.fullname = fullname;
+    }
 
-	public String getEmail() {
-		return email;
-	}
+    public String getEmail() {
+        return email;
+    }
 
-	public void setEmail(String email) {
-		this.email = email;
-	}
+    public void setEmail(String email) {
+        this.email = email;
+    }
 
-	public int getStatusaccount() {
-		return status;
-	}
+    public Long getStatusaccount() {
+        return status;
+    }
 
-	public void setStatusaccount(int statusaccount) {
-		this.status = statusaccount;
-	}
+    public void setStatusaccount(Long statusaccount) {
+        this.status = statusaccount;
+    }
 
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return authorities;
-	}
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
 
-	public void setAuthorities(Collection<? extends GrantedAuthority> authorities) {
-		this.authorities = authorities;
-	}
+    public void setAuthorities(Collection<? extends GrantedAuthority> authorities) {
+        this.authorities = authorities;
+    }
 
-	public static long getSerialversionuid() {
-		return serialVersionUID;
-	}
+    public static long getSerialversionuid() {
+        return serialVersionUID;
+    }
 
-	@Override
-	public boolean equals(Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-		UserDetailsImpl user = (UserDetailsImpl) o;
-		return Objects.equals(id, user.id);
-	}
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
+        UserDetailsImpl user = (UserDetailsImpl) o;
+        return Objects.equals(id, user.id);
+    }
 
-	@Override
-	public int hashCode() {
-		return super.hashCode();
-	}
+    @Override
+    public int hashCode() {
+        return super.hashCode();
+    }
 
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
 
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
 
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
 
-	@Override
-	public boolean isEnabled() {
-		boolean statusAccount = false;
-		if (status == 2) {
-			statusAccount = true;
-		}
-		return statusAccount;
-	}
+    @Override
+    public boolean isEnabled() {
+        boolean statusAccount = false;
+        if (status == 2) {
+            statusAccount = true;
+        }
+        return statusAccount;
+    }
+
 }
